test(plugin): update skill-md assertions for native-flow SKILL.md
